adding CPP Makefile for the megaphone ex00

+++ ex00/megaphone.cpp
@@ -6,7 +6,7 @@
 /*   By: azgaoua <azgaoua@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/24 02:24:17 by azgaoua           #+#    #+#             */
-/*   Updated: 2023/12/24 03:03:40 by azgaoua          ###   ########.fr       */
+/*   Updated: 2023/12/24 03:20:11 by azgaoua          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ int main(int ac, char **av)
 		default:
 			for (int k = 1; k < ac; k++) {
 				std::string str(av[k]);
-				for (int i = 0; i < str.length(); i++) {
+				for (size_t i = 0; i < str.length(); i++) {
 					str[i] = toupper(str[i]);
 				}
 				std::cout << str;
